Perceptron classifies if rock or metal cylinder
This is a project for my Intro to Neural Networks class. I Implemented a perceptron and used it on a set of sonar data to see how accurately I can get my perceptron to distinguish between two different classes of objects: rocks vs. metal cylinders. Once I implemented my perceptron, I trained it using a subset of inputs from the dataset, then validated it using the entire dataset. My program keeps track of how many inputs are correctly identified in order to calculate the accuracy of my perceptron. A test harness was also implemented that reads the data, creates the training set, initializes the perceptron, trains it and then tests it.